CurrentUserProfile - moved the formhandler outside of the xml segment of the file
Changed how errors are going to be handled throughout the applicaiton =
index.ts - changed the ErrorRequesthandler. Moved the overall logic outside of the file and into it's own called 'mainErrorHandler.ts'
Changed the name of the file 'nullChekcer.ts' to 'rowChecker.ts' to allign more with what it does, check if the row is null or not from a prisma retrieval.
Created a new class, ExpressError, in the file ExpressError.ts; extends the Error class and is meant to be primarily used as it's alternative throughout the application.